Moved controllers module one level up
