+++ morpheus/stages/input/rss_source_stage.py
@@ -19,11 +19,11 @@ import mrc
 
 from morpheus.cli import register_stage
 from morpheus.config import Config
+from morpheus.controllers.rss_controller import RSSController
 from morpheus.messages import MessageMeta
 from morpheus.pipeline.preallocator_mixin import PreallocatorMixin
 from morpheus.pipeline.single_output_source import SingleOutputSource
 from morpheus.pipeline.stream_pair import StreamPair
-from morpheus.utils.controllers.rss_controller import RSSController
 
 logger = logging.getLogger(__name__)
